Update manifest.json
Hi,

I am a Cloud Engineer in the Oracle Cloud Engineering Org. We are running this workshop as a webinar for OCI Customers as part of our customer outreach program. To be able to create an event code for this workshop and run it on live lab tenancy, Oracle Live Lab and Workshop Management Team has requested us to modify the "Getting Started" page URL in the manifest.json file. I have made the necessary changes and submitted this request. 

Please approve this request as soon as possible as we have upcoming events for this workshop in coming weeks. if you have any questions, please reach out to me via email/slack at abdul.rafae.mohammed@oracle.com.

Thanks!